Refactors the test to fail in new and exciting ways
