+++ test_fixtures/integrations/pytorch_lightning/__init__.py
@@ -76,19 +76,3 @@ class GenerateDataModule(LightningDataModule):
             batch_size=self.eval_batch_size,
             shuffle=self.eval_shuffle,
         )
-
-
-# TODO: remove replicated code
-@Step.register("generate_data")
-class GenerateData(Step):
-    DETERMINISTIC = True
-    CACHEABLE = False
-
-    def run(self) -> DatasetDict:  # type: ignore[override]
-        torch.manual_seed(1)
-        return DatasetDict(
-            {
-                "train": [{"x": torch.rand(10), "y": torch.rand(1)} for _ in range(64)],
-                "validation": [{"x": torch.rand(10), "y": torch.rand(1)} for _ in range(32)],
-            }
-        )
